requirements.txt
beautifulsoup4
click
lxml
numpy
pandas
python-dotenv
selenium